# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Unicis Platform API
  description: |
    REST API for the Unicis GRC (Governance, Risk & Compliance) platform.

    ## Authentication
    All endpoints require authentication via one of:
    - **Bearer token** — API key from team settings (`Authorization: Bearer <key>`)
    - **Session cookie** — browser session via NextAuth

    ## Response format
    Every response uses a consistent envelope:
    ```json
    { "data": <payload>, "error": null }       // success
    { "error": { "message": "..." } }          // error
    ```

    ## Roles & Permissions
    API keys authenticate with **ADMIN** role permissions.
    Available roles: `OWNER`, `ADMIN`, `MEMBER`, `AUDITOR`.
  version: 0.2.2
  contact:
    name: Unicis Tech
    url: https://unicis.tech
servers:
  - url: '{baseUrl}'
    variables:
      baseUrl:
        default: 'http://localhost:4002'
        description: Base URL of the Unicis Platform instance
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from team settings
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie (browser-based)
  parameters:
    slug:
      name: slug
      in: path
      required: true
      description: Team slug identifier
      schema:
        type: string
        example: my-team
    taskNumber:
      name: taskNumber
      in: path
      required: true
      description: Task number within the team
      schema:
        type: integer
        example: 1
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
          required:
            - message
    EmptyDataResponse:
      type: object
      properties:
        data:
          type: object
        error:
          nullable: true
    Task:
      type: object
      properties:
        id:
          type: integer
        taskNumber:
          type: integer
        teamId:
          type: string
        authorId:
          type: string
        title:
          type: string
        status:
          type: string
        priority:
          $ref: '#/components/schemas/TaskPriority'
        duedate:
          type: string
        description:
          type: string
        properties:
          type: object
          description: Flexible JSON properties (controls, risk, assessments, etc.)
        comments:
          type: array
          deprecated: true
          description: Legacy embedded comments. Use the dedicated task comments endpoint instead.
          items:
            $ref: '#/components/schemas/Comment'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
    TaskPriority:
      type: string
      enum:
        - low
        - medium
        - high
    Comment:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        taskId:
          type: integer
        createdById:
          type: string
        createdBy:
          type: object
          properties:
            name:
              type: string
              nullable: true
            email:
              type: string
...
    Attachment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: integer
        filename:
          type: string
        url:
          type: string
...
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
          description: Task title
        status:
          type: string
          description: Task status
        priority:
          allOf:
            - $ref: '#/components/schemas/TaskPriority'
          description: Task priority. Defaults to medium when omitted.
        duedate:
          type: string
          description: Due date
        description:
          type: string
          description: Rich text description (HTML sanitized)
      required:
        - title
        - status
        - duedate
    UpdateTaskRequest:
      type: object
      properties:
        data:
          type: object
          description: Fields to update
          properties:
            title:
              type: string
            status:
              type: string
            priority:
              $ref: '#/components/schemas/TaskPriority'
            duedate:
              type: string
            description:
              type: string
          additionalProperties: true
      required:
        - data
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              message: Unauthorized
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              message: You are not allowed to perform this action
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              message: Not found
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /api/teams/{slug}/tasks:
    parameters:
      - $ref: '#/components/parameters/slug'

    get:
      operationId: listTasks
      tags: [Tasks]
      summary: List all tasks for a team
      description: Returns all tasks belonging to the specified team. Requires `task:read` permission.
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  error:
                    nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      operationId: createTask
      tags: [Tasks]
      summary: Create a new task
      description: Creates a new task in the team. Requires `task:create` permission. The description field is sanitized for XSS.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '200':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyDataResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/teams/{slug}/tasks/{taskNumber}:
    parameters:
      - $ref: '#/components/parameters/slug'
      - $ref: '#/components/parameters/taskNumber'

    get:
      operationId: getTask
      tags: [Tasks]
      summary: Get a single task
      description: Returns a task by its number within the team. Requires `task:read` permission. Embedded comments are kept for backwards compatibility and can be omitted.
      parameters:
        - name: includeComments
          in: query
          required: false
          description: Set to false when comments are loaded through the dedicated comments endpoint.
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
                  error:
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateTask
      tags: [Tasks]
      summary: Update a task
      description: Updates task fields. Requires `task:update` permission. Description is sanitized for XSS.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Updated task
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
                  error:
                    nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteTask
      tags: [Tasks]
      summary: Delete a task
      description: Permanently deletes a task. Requires `task:delete` permission.
      responses:
        '200':
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyDataResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...